feat(infra): add support for multiple isolated environments
Terraform 0.13 is now required.

This allows separate Terraform setups for staging and production
environments for example, while the core resources and config lives
within modules which are used by all environments.

It also improves the Terraform setup in general, using `for_each`
instead of `count` to iterate over resources, which means that Terraform
can refer to each resource by name, rather than a list index
offset. Previously if a named host in the middle of the list was
removed, it would taint all resources after it, as their index offset
was now different.

The Terraform 0.13 upgrade also involves a switch to community managed
providers. This means the existing state file in production will have
resources belonging to providers which no longer exist, and will error
out. To be able to apply this to production, the state file needs to
first be updated, changing the providers over to the new FQDNs. The
following commands should do so:

    terraform state replace-provider \
      registry.terraform.io/terraform-providers/cloudflare \
      registry.terraform.io/cloudflare/cloudflare
    terraform state replace-provider \
      registry.terraform.io/terraform-providers/scaleway \
      registry.terraform.io/scaleway/scaleway

Usage now involves running Terraform from one of the environment
specific workspace directories:

- infrastructure/production
- infrastructure/staging

Ansible is now designed to by executed by Terraform, rather than
directly running it. This ensures that it is properly configured with
all required and relevant configuration. Terraform will always execute
Ansible at the end of a `terraform apply` run. To skip ansible, set the
`ansible_skip` Terraform variable to `true` with one of the following:

- `terraform apply -var 'ansible_skip=true`
- `TF_VAR_ansible_skip=true terraform apply`

It will still mark the `module.ansible.null_resource.run` as needing a
change, but with skip set to true, it will not execute
`ansible-playbook`.

Check `infrastructure/<environment>/variables.tf` files for more details
about available variables.