chore(infra): make reverse DNS optional and disabled by default
It seems Scaleway's backend systems are sometimes slow to pickup DNS
changes in cloudflare, and refuse to accept reverse DNS changes unless
they believe the name resolves to the IP in question.

Due to this, we for now disable reverse DNS by default, until we can
come up with a more reliable solution.