feat(internal/net): add object buffer on connections
Buffer allows us to hold the last x number of objects received by
this connection and replay them when a new subscription is created.
This is done as subscriptions some times are not created fast enough
after a new connection is created and the subscribers are losing
objects. This is especially annoying in tests.