refactor(identity): rename IdentityID to Identity, and Identity to KeyGraph
after a lengthly chat with @jimeh, we kinda settled at some slight renames to improve the consistency of terms that the backend and various frontends will use. The identifier for an entity (person/company/etc) is now the Identity which points to a KeyGraph. The name KeyGraph was preferred to KeyStream as the next part of this saga will be to re-work the names of Documents and Streams, and it seems likely the term Stream is going away.